Marge working directory
