+++ .gitignore
@@ -27,7 +27,6 @@ _testmain.go
 *.test
 *.prof
 
-/apig
 /bin/
 
 /bindata.go
