+++ apig/generate_test.go
@@ -187,3 +187,29 @@ func TestGenerateRouter(t *testing.T) {
 		t.Fatalf("Failed to generate router correctly.\nexpected:\n%s\nactual:\n%s", string(c1), string(c2))
 	}
 }
+
+func TestGenerateDB(t *testing.T) {
+	outDir, err := ioutil.TempDir("", "generateDB")
+	if err != nil {
+		t.Fatal("Failed to create tempdir")
+	}
+	defer os.RemoveAll(outDir)
+
+	if err := generateDB(detail, outDir); err != nil {
+		t.Fatalf("Error should not be raised: %s", err)
+	}
+
+	path := filepath.Join(outDir, "db", "db.go")
+	_, err = os.Stat(path)
+	if err != nil {
+		t.Fatalf("Router file is not generated: %s", path)
+	}
+
+	fixture := filepath.Join("testdata", "db", "db.go")
+
+	if !compareFiles(path, fixture) {
+		c1, _ := ioutil.ReadFile(fixture)
+		c2, _ := ioutil.ReadFile(path)
+		t.Fatalf("Failed to generate db.go correctly.\nexpected:\n%s\nactual:\n%s", string(c1), string(c2))
+	}
+}
