+++ command/gen.go
