+++ command/gen_test.go
