+++ command/meta.go
