+++ command/new.go
