--- tmp/README.md
@@ -1,29 +1,0 @@
-# apig
-
-
-
-## Description
-
-## Usage
-
-## Install
-
-To install, use `go get`:
-
-```bash
-$ go get -d github.com/wantedly/apig
-```
-
-## Contribution
-
-1. Fork ([https://github.com/wantedly/apig/fork](https://github.com/wantedly/apig/fork))
-1. Create a feature branch
-1. Commit your changes
-1. Rebase your local changes against the master branch
-1. Run test suite with the `go test ./...` command and confirm that it passes
-1. Run `gofmt -s`
-1. Create a new Pull Request
-
-## Author
-
-[wantedly](https://github.com/wantedly)
