--- tmp/apig/testdata/README.md
@@ -1,17 +1,0 @@
-# API Server
-
-Simple Rest API using gin(framework) & gorm(orm)
-
-## Endpoint list
-
-### Users Resource
-
-```
-GET    /api/users
-GET    /api/users/:id
-POST   /api/users
-PUT    /api/users/:id
-DELETE /api/users/:id
-```
-
-server runs at http://localhost:8080
