--- tmp/apig/testdata/docs/index.apib
@@ -1,6 +1,0 @@
-FORMAT: 1A
-HOST: http://localhost:8080
-
-# Api-Server API
-
-<!-- include(user.apib) -->
