--- tmp/apig/testdata/models.go
@@ -1,20 +1,0 @@
-package model
-
-import (
-	"time"
-)
-
-type User struct {
-	ID        uint       `gorm:"primary_key;AUTO_INCREMENT" json:"id,omitempty"`
-	Name      string     `json:"name,omitempty"`
-	CreatedAt *time.Time `json:"created_at,omitempty"`
-	UpdatedAt *time.Time `json:"updated_at,omitempty"`
-}
-
-type Job struct {
-	ID          uint       `gorm:"primary_key;AUTO_INCREMENT" json:"id,omitempty"`
-	Name        string     `json:"name,omitempty"`
-	Description string     `json:"description,omitempty"`
-	CreatedAt   *time.Time `json:"created_at,omitempty"`
-	UpdatedAt   *time.Time `json:"updated_at,omitempty"`
-}
